requirements.txt
lightning
hydra-core
pandas
matplotlib
seaborn
jupyterlab
ipykernel
wandb
numpy
huggingface_hub
python-dotenv